Simplify and correct concurrency
By moving the creation and closing of intermediate channels to the noDuplicates
method, the recursion could be made more explicit. Once that was simplified, it
became clear that the generator channel and the head channel were redundant as
well, and so the for-select loop was down to a single channel. After removing
that extra code, there are fewer races, the set of codes is populated, and the
channels are all cascaded closed (we could wait for the next channel to drain,
but there is no real need yet).